added new flame thrower and animated creeps

+++ src/main/java/lineup/model/implementations/projectiles/LightRocket.java
@@ -47,7 +47,7 @@ public class LightRocket extends Projectile {
   }
 
   @Override
-  public int getDamage() {
+  public double getDamage() {
     return 5;
   }
