+++ src/main/java/lineup/model/implementations/projectiles/MediumBullet.java
@@ -39,7 +39,7 @@ public class MediumBullet extends Projectile {
   }
 
   @Override
-  public int getDamage() {
+  public double getDamage() {
     return 2;
   }
